+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Release Notes for Product Labels
 
 ## [Unreleased]
+### Added
+- Added Product Types condition rule for matching products by product types (#2)
+
+[#1]: https://github.com/thepixelage/craft-productlabels/issues/1
 
 ## 4.0.1 - 2022-05-05
 ### Fixed
 - Fixed Craft CMS and Craft Commerce dependency versions in composer.json
-
 
 ## 4.0.0 - 2022-05-05
 ### Added